Refactor DataSourceConfiguration
