+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/config/datasource/creator/impl/DefaultDataSourceCreator.java
@@ -1,0 +1,50 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shardingsphere.infra.config.datasource.creator.impl;
+
+import org.apache.shardingsphere.infra.config.datasource.DataSourceConfiguration;
+
+import javax.sql.DataSource;
+import java.lang.reflect.Method;
+import java.util.Map;
+
+/**
+ * Default data source creator.
+ */
+public final class DefaultDataSourceCreator extends AbstractDataSourceCreator {
+    
+    @Override
+    public DataSourceConfiguration createDataSourceConfiguration(final DataSource dataSource) {
+        return buildDataSourceConfiguration(dataSource);
+    }
+    
+    @Override
+    public DataSource createDataSource(final DataSourceConfiguration dataSourceConfiguration) {
+        DataSource result = buildDataSource(dataSourceConfiguration.getDataSourceClassName());
+        Method[] methods = result.getClass().getMethods();
+        for (Map.Entry<String, Object> entry : dataSourceConfiguration.getAllProps().entrySet()) {
+            setField(result, methods, entry.getKey(), entry.getValue());
+        }
+        return result;
+    }
+    
+    @Override
+    public String getType() {
+        return "";
+    }
+}
