+++ shardingsphere-mode/shardingsphere-mode-type/shardingsphere-cluster-mode/shardingsphere-cluster-mode-core/src/test/java/org/apache/shardingsphere/mode/manager/cluster/ClusterContextManagerBuilderTest.java
@@ -23,6 +23,7 @@ import org.apache.shardingsphere.authority.rule.AuthorityRule;
 import org.apache.shardingsphere.infra.config.RuleConfiguration;
 import org.apache.shardingsphere.infra.config.algorithm.ShardingSphereAlgorithmConfiguration;
 import org.apache.shardingsphere.infra.config.datasource.DataSourceConfiguration;
+import org.apache.shardingsphere.infra.config.datasource.DataSourceConverter;
 import org.apache.shardingsphere.infra.config.mode.ModeConfiguration;
 import org.apache.shardingsphere.infra.config.mode.PersistRepositoryConfiguration;
 import org.apache.shardingsphere.infra.config.properties.ConfigurationProperties;
@@ -136,9 +137,9 @@ public final class ClusterContextManagerBuilderTest {
     private Map<String, DataSourceConfiguration> getDataSourceConfigurations() {
         MockedDataSource dataSource = new MockedDataSource();
         Map<String, DataSourceConfiguration> result = new LinkedHashMap<>(3, 1);
-        result.put("primary_ds", DataSourceConfiguration.getDataSourceConfiguration(dataSource));
-        result.put("ds_0", DataSourceConfiguration.getDataSourceConfiguration(dataSource));
-        result.put("ds_1", DataSourceConfiguration.getDataSourceConfiguration(dataSource));
+        result.put("primary_ds", DataSourceConverter.getDataSourceConfiguration(dataSource));
+        result.put("ds_0", DataSourceConverter.getDataSourceConfiguration(dataSource));
+        result.put("ds_1", DataSourceConverter.getDataSourceConfiguration(dataSource));
         return result;
     }
     
@@ -197,9 +198,9 @@ public final class ClusterContextManagerBuilderTest {
     private Map<String, DataSourceConfiguration> getChangedDataSourceConfigurations() {
         MockedDataSource dataSource = new MockedDataSource();
         Map<String, DataSourceConfiguration> result = new LinkedHashMap<>(3, 1);
-        result.put("primary_ds", DataSourceConfiguration.getDataSourceConfiguration(dataSource));
-        result.put("ds_1", DataSourceConfiguration.getDataSourceConfiguration(dataSource));
-        result.put("ds_2", DataSourceConfiguration.getDataSourceConfiguration(dataSource));
+        result.put("primary_ds", DataSourceConverter.getDataSourceConfiguration(dataSource));
+        result.put("ds_1", DataSourceConverter.getDataSourceConfiguration(dataSource));
+        result.put("ds_2", DataSourceConverter.getDataSourceConfiguration(dataSource));
         return result;
     }
